+++ example/src/services/bundler/bundler.worker.ts
@@ -1,4 +1,5 @@
-import { Volume } from "memfs";
+import { fs, InMemory } from "@zenfs/core";
+import { dirname } from "path";
 import dayjs from "dayjs";
 import esbuildWasmURL from "esbuild-wasm/esbuild.wasm?url";
 import {
@@ -11,10 +12,10 @@ import installTailwindPlugin from "@bundler-in-browser/tailwindcss";
 import type { PartialMessage } from "esbuild-wasm";
 import { fsData } from "../../fsData";
 import {
-  isCompilationRequest,
+  isBuildRequest,
   newWorkerReadyMessage,
   workerSourceMarker,
-  type CompilationSuccessResponse,
+  type BuildSuccessResponse,
   type NPMProgressMessage,
   type WorkerLogMessage,
 } from "./common";
@@ -24,7 +25,10 @@ main().catch((err) => {
   self.postMessage(newWorkerReadyMessage(err));
 });
 async function main() {
-  const fs = Volume.fromJSON(fsData);
+  Object.entries(fsData).forEach(([path, content]) => {
+    fs.mkdirSync(dirname(path), { recursive: true });
+    fs.writeFileSync(path, content)
+  });
   const bundler = new BundlerInBrowser(fs);
 
   // setup event listeners
@@ -33,15 +37,22 @@ async function main() {
   bundler.events.on("npm:progress", (e) => log(`[npm] [${e.stage}] [${e.current} / ${e.total}] ${e.packageId}`));
   bundler.events.on("npm:install:done", () => log(`[npm] install:done`));
   bundler.events.on("npm:install:error", (e) => log(`[npm] install:error`, e.errors));
-  bundler.events.on("compile:start", () => log("compile:start"));
-  bundler.events.on("compile:usercode", (result) => log("compile:usercode", result));
-  bundler.events.on("compile:vendor", (result) => log("compile:vendor", result));
+  bundler.events.on("build:start", () => log("build:start"));
+  bundler.events.on("build:usercode", (result) => log("build:usercode", result));
+  bundler.events.on("build:vendor", (result) => log("build:vendor", result));
 
   bundler.events.on("npm:install:error", () => self.postMessage({ type: 'npm-progress', source: workerSourceMarker, event: 'error', progress: null } satisfies NPMProgressMessage));
   bundler.events.on("npm:install:done", () => self.postMessage({ type: 'npm-progress', source: workerSourceMarker, event: 'done', progress: null } satisfies NPMProgressMessage));
   bundler.events.on("npm:progress", (e) => self.postMessage({ type: 'npm-progress', source: workerSourceMarker, event: 'progress', progress: e } satisfies NPMProgressMessage));
 
   // initialize
+
+  bundler.npm.options.blocklist = [
+    '@vue/compiler-core',
+    '@vue/compiler-dom',
+    '@vue/compiler-sfc',
+    '@vue/server-renderer',
+  ]
 
   await bundler.initialize({
     esbuildWasmURL: esbuildWasmURL,
@@ -59,30 +70,37 @@ async function main() {
 
   // ready to compile
   self.addEventListener("message", async ({ data }) => {
-    if (!isCompilationRequest(data)) return;
-    fs.fromJSON(data.files);
+    if (!isBuildRequest(data)) return;
+
+    // reset fs
+    fs.umount('/')
+    fs.mount('/', InMemory.create({}));
+    Object.entries(data.files).forEach(([path, content]) => {
+      fs.mkdirSync(dirname(path), { recursive: true });
+      fs.writeFileSync(path, content);
+    });
 
     await bundler
-      .compile({
+      .build({
         entrypoint: '/src/index.js',
       })
       .then((result) => {
-        const message: CompilationSuccessResponse = {
+        const message: BuildSuccessResponse = {
           result,
           wrappedJs: wrapCommonJS(result.js),
         };
         data.port.postMessage(message);
-        log("compile:done", result);
+        log("build:done", result);
       })
       .catch((err: Error & { errors?: (PartialMessage | Error)[] }) => {
         data.port.postMessage({ errors: err.errors || [err] });
 
-        log(`compile:error ${err}`);
+        log(`build:error ${err}`);
         // if (err.errors?.length) {
         //   err.errors.forEach((error, index, errors) => {
         //     let msg = 'text' in error ? error.text : String(error);
         //     let pos = 'location' in error && error.location ? ` (${error.location.file}:${error.location.line}:${error.location.column})` : '';
-        //     log(`compile:error ${index + 1}/${errors.length}`, msg, pos);
+        //     log(`build:error ${index + 1}/${errors.length}`, msg, pos);
         //   })
         // } else {
         //   log(err);